refactor(cli): split installer into cohesive modules
